Restructures project & moves fai components (#24)
* Adds project structure for fai pipeline

- goal is to create a create a stand alone directory containing all fai components
- uses the __main__ naming convention; this may be incorrect given the current project structure

* Moves fai operations to new fai pipeline scripts

- moves fai components from __main__ to pipelines/fast_ai_1/__main__
- extraction of a DataBunch df was left in __main__ as it could be seen as
  a useful general procedure (still tbd)
- moves fai helpers from helpers/utilities to pipelines/fast_ai_1/helpers/utilities

* Moves fastai_models to "pipelines.fast_ai_1.model"
