+++ src/__main__.py
@@ -1,20 +1,19 @@
 import argparse
+import json
 import logging.config
 from ast import literal_eval
 from os import path
 
-from fastai.vision import *
-
 from src.data.data_model import *
-from src.helpers.utilities import *
 from src.eda.describe_leaf import plot_embolisms_per_leaf
-from src.model.fastai_models import FastaiUnetLearner
+from src.helpers.utilities import *
 
 abs_path = path.dirname(path.abspath(__file__))
 
 logging.config.fileConfig(fname=abs_path + "/logging_configuration.ini",
                           defaults={'logfilename': abs_path + "/main.log"},
                           disable_existing_loggers=False)
+
 LOGGER = logging.getLogger(__name__)
 
 pil_logger = logging.getLogger('PIL')
@@ -167,7 +166,7 @@ def plot_mseq_embolism_counts(mseqs, show, output_path, tiles,
 
                     mask_image.load_tile_paths()
                     mask_image.load_extracted_images(load_image=True,
-                                                    disable_pb=True)
+                                                     disable_pb=True)
                     mask_image.get_embolism_percent_list(disable_pb=True)
                     mask_image.get_has_embolism_list(disable_pb=True)
                     has_embolism_lol[i] = \
@@ -249,48 +248,6 @@ def extract_tiles_databunch_df(lseqs, mseqs, output_path_list,
         _ = lseq.get_tile_databunch_df(mseq, tile_embolism_only,
                                        leaf_embolism_only,
                                        csv_output_path)
-
-
-def train_fastai_unet(train_df_paths, save_path, bs, epochs, lr,
-                      val_df_paths=None, unfreeze_type=None,
-                      model_weights_path=None, **kwargs):
-    training_dfs = [pd.read_csv(path, index_col=0) for path in train_df_paths]
-
-    if val_df_paths:
-        validation_dfs = [pd.read_csv(path, index_col=0) for path in val_df_paths]
-        combined_df = combine_and_add_valid(training_dfs, validation_dfs)
-    else:
-        combined_df = pd.concat(training_dfs)
-
-    combined_df, folder_name_path = format_databunch_df(
-        combined_df, "folder_path", "leaf_name", create_copy=True)
-
-    fai = FastaiUnetLearner()
-
-    if val_df_paths:
-        fai.prep_fastai_data(combined_df, folder_name_path, bs, plot=False,
-                             mask_col_name="mask_path",
-                             **kwargs)
-    else:
-        fai.prep_fastai_data(combined_df, folder_name_path, bs, plot=False,
-                             mask_col_name="mask_path",
-                             split_func=ItemList.split_by_rand_pct,
-                             **kwargs)
-
-    fai.create_learner()
-
-    if model_weights_path:
-        fai.load_weights(model_weights_path)
-
-    fai.train(epochs=epochs, save_path=save_path, lr=lr,
-              unfreeze_type=unfreeze_type)
-
-
-def predict_fastai_unet(lseqs, length_x, length_y, model_pkl_path):
-    fai_unet_learner = FastaiUnetLearner(model_pkl_path=model_pkl_path)
-
-    for lseq in lseqs:
-        lseq.predict_leaf_sequence(fai_unet_learner, length_x, length_y)
 
 
 def parse_arguments() -> argparse.Namespace:
@@ -386,7 +343,7 @@ def parse_arguments() -> argparse.Namespace:
         "--show", "-s", action="store_true", help="flag indicating if the "
                                                   "plot should be shown")
     parser_plot_profile.add_argument(
-        "--leaf_names", "-ln",  type=str, metavar="\b",
+        "--leaf_names", "-ln", type=str, metavar="\b",
         help="leaf names to be used in plot title")
 
     parser_plot_embolism_counts = subparsers.add_parser(
@@ -400,7 +357,7 @@ def parse_arguments() -> argparse.Namespace:
         "--show", "-s", action="store_true", help="flag indicating if the "
                                                   "plot should be shown")
     parser_plot_embolism_counts.add_argument(
-        "--leaf_names", "-ln",  type=str, metavar="\b",
+        "--leaf_names", "-ln", type=str, metavar="\b",
         help="leaf names to be used in plot title")
     parser_plot_embolism_counts.add_argument(
         "--tile", "-t", action="store_true",
@@ -437,41 +394,6 @@ def parse_arguments() -> argparse.Namespace:
     parser_databunch_df.add_argument(
         "--leaf_embolism_only", "-leo", action="store_true",
         help="should only full leafs with embolisms be used")
-
-    parser_train_fastai = subparsers.add_parser(
-        "train_fastai", help="train a fastai unet model")
-    parser_train_fastai.set_defaults(which="train_fastai")
-    parser_train_fastai.add_argument("--batch_size", "-bs", type=int,
-                                     metavar="\b", help="batch size")
-    parser_train_fastai.add_argument("--save_path", "-sp", type=str,
-                                     metavar="\b",
-                                     help="path to save model weights and "
-                                          "model pickle")
-    parser_train_fastai.add_argument("--epochs", "-e", type=int,
-                                     metavar="\b", help="number of epochs")
-    parser_train_fastai.add_argument("--learning_rate", "-lr", type=float,
-                                     metavar="\b", help="learning rate")
-    parser_train_fastai.add_argument(
-        "--unfreeze_type", "-ut", metavar="\b",
-        help="whether the model needs to be unfrozen before training, "
-             "the options that follow refer to the layer you want to "
-             "unfreeze: 'all', 'last', or the number you want to freeze up "
-             "to")
-    parser_train_fastai.add_argument(
-        "--model_weights_path", "-mwp",
-        help="path to fastai model weights (complete path excluding .pth) to "
-             "retrain, if the paths are in the input json enter \"same\"")
-
-    parser_predict_fastai = subparsers.add_parser(
-        "predict_fastai", help="train a fastai unet model")
-    parser_predict_fastai.set_defaults(which="predict_fastai")
-    parser_predict_fastai.add_argument("length_x", type=int,
-                                       help="tile x length")
-    parser_predict_fastai.add_argument("length_y", type=int,
-                                       help="tile y length")
-    parser_predict_fastai.add_argument(
-        "model_path", help="path to fastai pkl (complete path), if the paths "
-                           "are in the input json enter \"same\"")
 
     args = parser.parse_args()
     return args
@@ -639,40 +561,3 @@ if __name__ == "__main__":
         else:
             extract_full_databunch_df(
                 LSEQS, MSEQS, CSV_OUTPUT_LIST, ARGS.leaf_embolism_only)
-
-    if ARGS.which == "train_fastai":
-        TRAIN_CSV_INPUT_LIST = INPUT_JSON_DICT["fastai"]["train"]
-        VAL_CSV_INPUT_LIST = INPUT_JSON_DICT["fastai"]["validation"]
-
-        if ARGS.unfreeze_type:
-            if ARGS.unfreeze_type != "all" and ARGS.unfreeze_type != "last":
-                UNFREEZE_TYPE = int(ARGS.unfreeze_type)
-            else:
-                UNFREEZE_TYPE = ARGS.unfreeze_type
-        else:
-            UNFREEZE_TYPE = None
-
-        if VAL_CSV_INPUT_LIST:
-            train_fastai_unet(TRAIN_CSV_INPUT_LIST, save_path=ARGS.save_path,
-                              bs=ARGS.batch_size, epochs=ARGS.epochs,
-                              lr=ARGS.learning_rate,
-                              val_df_paths=VAL_CSV_INPUT_LIST,
-                              unfreeze_type=UNFREEZE_TYPE,
-                              model_weights_path=ARGS.model_weights_path)
-        else:
-            train_fastai_unet(TRAIN_CSV_INPUT_LIST,
-                              save_path=ARGS.save_path, bs=ARGS.batch_size,
-                              epochs=ARGS.epochs, lr=ARGS.learning_rate,
-                              unfreeze_type=UNFREEZE_TYPE,
-                              model_weights_path=ARGS.model_weights_path, seed=314)
-
-    if ARGS.which == "predict_fastai":
-        if ARGS.model_path == "same":
-            MODEL_PATH = INPUT_JSON_DICT["fastai"]["model"]
-        else:
-            MODEL_PATH = ARGS.model_path
-
-        load_image_objects(LSEQS)
-
-        predict_fastai_unet(LSEQS, ARGS.length_x, ARGS.length_y,
-                            MODEL_PATH)
